+++ example/src/main/java/org/blockstack/android/MainActivity.kt
@@ -49,6 +49,7 @@ class MainActivity : AppCompatActivity() {
         putImageFileButton.isEnabled = false
         getStringFileFromUserButton.isEnabled = false
         getAppBucketUrlButton.isEnabled = false
+        listFilesButton.isEnabled = false
 
 
         signInButton.setOnClickListener { _: View ->
@@ -212,6 +213,22 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
+        listFilesButton.setOnClickListener {
+            listFilesText.text = "...."
+            blockstackSession().listFiles ({urlResult ->
+                if (urlResult.hasValue) {
+                    if (listFilesText.text === "....") {
+                        listFilesText.text = urlResult.value
+                    } else {
+                        listFilesText.text = listFilesText.text.toString() + "\n" + urlResult.value
+                    }
+                }
+                true
+            }, { countResult ->
+                Log.d(TAG, "files count " + if (countResult.hasValue) {countResult.value} else {countResult.error})
+            })
+        }
+
         if (intent?.action == Intent.ACTION_VIEW) {
             handleAuthResponse(intent)
         }
@@ -230,6 +247,7 @@ class MainActivity : AppCompatActivity() {
         getImageFileButton.isEnabled = true
         getStringFileFromUserButton.isEnabled = true
         getAppBucketUrlButton.isEnabled = true
+        listFilesButton.isEnabled = true
     }
 
     private fun showUserAvatar(avatarImage: String?) {