+++ CHANGELOG.md
@@ -10,6 +10,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- `BlockstackSession.listFiles` iterates through the user's app files   
+- `Scope.fromJSName` creates `Scope` from its javascript name as used in blockstack.js
+ 
 ## [0.4.1] - 2018-11-05
 
 ### Changed